# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Castellan API
  version: 0.1.0
  description: |
    Usage-based API monetization gateway with Stellar settlement.

    Two authentication schemes are used:
      - **API keys** — prefixed `ca_`, long-lived, passed via
        `Authorization: Bearer ca_<key>`, used for server-to-server / gateway
      - **OAuth session** — dashboard users authenticate via Google or GitHub
        OAuth; the backend sets an httpOnly `session_token` cookie. No bearer
        header is needed for browser requests to the management API.
servers:
  - url: http://localhost:8080
    description: Local development
  - url: https://api.castellan.dev
    description: Production
paths:
...
  /api/v1/keys:
    get:
      operationId: listKeys
      summary: List all API keys for the authenticated consumer
      security: [{ BearerAuth: [] }, { CookieAuth: [] }]
      responses:
        "200":
          description: List of API keys
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ListKeysItem"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/InternalError"
    post:
      operationId: createKey
      summary: Create a new API key
      security: [{ BearerAuth: [] }, { CookieAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateKeyRequest"
      responses:
        "201":
          description: Created API key
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreateKeyResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/v1/keys/{id}/revoke:
    post:
      operationId: revokeKey
      summary: Revoke an API key
      security: [{ BearerAuth: [] }, { CookieAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: API key ID
      responses:
        "200":
          description: Revoked API key
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RevokeKeyResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/v1/keys/{id}/rotate:
    post:
      operationId: rotateKey
      summary: Rotate an API key (revoke old, return new)
      security: [{ BearerAuth: [] }, { CookieAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: API key ID
      responses:
        "200":
          description: New API key
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreateKeyResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"

  # ──────────────────────────────────────────────
  #   OAuth
  # ──────────────────────────────────────────────
  /auth/{provider}/login:
    get:
      operationId: oauthLogin
      summary: Initiate OAuth login
      description: |
        Redirects the user to the OAuth provider (google or github) for
        authentication. On success the provider redirects to the callback
        endpoint.
      parameters:
        - name: provider
          in: path
          required: true
          schema:
            type: string
            enum: [google, github]
          description: OAuth provider
      security: []
      responses:
        "302":
          description: Redirect to OAuth provider
        "400":
          description: Unsupported provider
  /auth/{provider}/callback:
    get:
      operationId: oauthCallback
      summary: OAuth callback
      description: |
        Handles the OAuth provider callback — exchanges the authorization
        code for user info, upserts the user by email, creates a session,
        sets the `session_token` cookie, and redirects to the dashboard.
      parameters:
        - name: provider
          in: path
          required: true
          schema:
            type: string
            enum: [google, github]
          description: OAuth provider
      security: []
      responses:
        "302":
          description: Redirect to dashboard after successful login
        "401":
          description: Authentication failed
        "500":
          description: Internal error

  # ──────────────────────────────────────────────
  #   Auth (session management)
  # ──────────────────────────────────────────────
  /api/v1/auth/me:
    get:
      operationId: getMe
      summary: Get the current user
      description: |
        Returns the authenticated user's profile. The session is read from
        the `session_token` cookie — no Authorization header is needed for
        browser requests.
      security: [{ CookieAuth: [] }]
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/v1/auth/logout:
    post:
      operationId: logout
      summary: Log the current user out
      description: |
        Revokes the session and clears the `session_token` cookie.
      security: [{ CookieAuth: [] }]
      responses:
        "200":
          description: Logged out
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: logged out
        "401":
          $ref: "#/components/responses/Unauthorized"

  # ──────────────────────────────────────────────
  #   Providers
  # ──────────────────────────────────────────────
  /api/v1/providers:
    post:
      operationId: createProvider
      summary: Register a new upstream provider
      security: [{ BearerAuth: [] }, { CookieAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProviderRequest"
      responses:
        "201":
          description: Created provider
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Provider"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/InternalError"
    get:
      operationId: listProviders
      summary: List all providers for the authenticated consumer
      security: [{ BearerAuth: [] }, { CookieAuth: [] }]
      responses:
        "200":
          description: List of providers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Provider"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/v1/providers/{id}:
    get:
      operationId: getProvider
      summary: Get a provider by ID
      security: [{ BearerAuth: [] }, { CookieAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: Provider ID
      responses:
        "200":
          description: Provider details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Provider"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
    patch:
      operationId: updateProvider
      summary: Partially update a provider's name or base URL
      security: [{ BearerAuth: [] }, { CookieAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: Provider ID
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProviderRequest"
      responses:
        "200":
          description: Updated provider
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Provider"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/v1/providers/{id}/status:
    patch:
      operationId: updateProviderStatus
      summary: Update a provider's status
      security: [{ BearerAuth: [] }, { CookieAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: Provider ID
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProviderStatusRequest"
      responses:
        "200":
          description: Updated provider
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Provider"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
    delete:
      operationId: deleteProvider
      summary: Delete a provider and all its endpoints
      security: [{ BearerAuth: [] }, { CookieAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: Provider ID
      responses:
        "204":
          description: Provider deleted (no content)
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"

  # ──────────────────────────────────────────────
  #   Endpoints
  # ──────────────────────────────────────────────
  /api/v1/providers/{providerId}/endpoints:
    post:
      operationId: createEndpoint
      summary: Register a new endpoint under a provider
      security: [{ BearerAuth: [] }, { CookieAuth: [] }]
      parameters:
        - name: providerId
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: Provider ID
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateEndpointRequest"
      responses:
        "201":
          description: Created endpoint
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Endpoint"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        "500":
          $ref: "#/components/responses/InternalError"
    get:
      operationId: listEndpoints
      summary: List all endpoints for a provider
      security: [{ BearerAuth: [] }, { CookieAuth: [] }]
      parameters:
        - name: providerId
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: Provider ID
        - name: status
          in: query
          required: false
          schema:
            type: string
            enum: [active, inactive, draft]
          description: Filter by endpoint status
      responses:
        "200":
          description: List of endpoints
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Endpoint"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/v1/endpoints/{id}:
    get:
      operationId: getEndpoint
      summary: Get an endpoint by ID
      security: [{ BearerAuth: [] }, { CookieAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: Endpoint ID
      responses:
        "200":
          description: Endpoint details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Endpoint"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
    patch:
      operationId: updateEndpoint
      summary: Partially update an endpoint's configuration
      security: [{ BearerAuth: [] }, { CookieAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: Endpoint ID
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateEndpointRequest"
      responses:
        "200":
          description: Updated endpoint
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Endpoint"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        "500":
          $ref: "#/components/responses/InternalError"
    delete:
      operationId: deleteEndpoint
      summary: Delete an endpoint
      security: [{ BearerAuth: [] }, { CookieAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: Endpoint ID
      responses:
        "204":
          description: Endpoint deleted (no content)
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/v1/endpoints/{id}/status:
    patch:
      operationId: updateEndpointStatus
      summary: Update an endpoint's status
      security: [{ BearerAuth: [] }, { CookieAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: Endpoint ID
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateEndpointStatusRequest"
      responses:
        "200":
          description: Updated endpoint
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Endpoint"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"

  # ──────────────────────────────────────────────
  #   Accounts
  # ──────────────────────────────────────────────
  /api/v1/accounts/me:
    get:
      operationId: getAccount
      summary: Get the authenticated consumer's account
      security: [{ BearerAuth: [] }, { CookieAuth: [] }]
      responses:
        "200":
          description: Account details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AccountResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/v1/accounts/me/entries:
    get:
      operationId: listEntries
      summary: List ledger entries for the authenticated consumer
      security: [{ BearerAuth: [] }, { CookieAuth: [] }]
      parameters:
        - name: type
          in: query
          required: false
          schema:
            type: string
            enum: [deposit, reservation, deduction, refund, settlement]
          description: Filter by entry type
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 50
          description: Maximum results (max 100)
        - name: offset
          in: query
          required: false
          schema:
            type: integer
            minimum: 0
            default: 0
          description: Number of results to skip
      responses:
        "200":
          description: Paginated list of entries
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ListEntriesResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/v1/accounts/me/entries/{id}:
    get:
      operationId: getEntry
      summary: Get a specific ledger entry
      security: [{ BearerAuth: [] }, { CookieAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: Entry ID
      responses:
        "200":
          description: Entry details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EntryResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"

  # ──────────────────────────────────────────────
  #   Deposits
  # ──────────────────────────────────────────────
  /api/v1/deposits/intent:
    get:
      operationId: depositIntent
      summary: Get deposit instructions (SEP-7 URI + QR code)
      security: [{ BearerAuth: [] }, { CookieAuth: [] }]
      responses:
        "200":
          description: Deposit intent details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DepositIntentResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/v1/deposits:
    get:
      operationId: listDeposits
      summary: List deposit history
      security: [{ BearerAuth: [] }, { CookieAuth: [] }]
      responses:
        "200":
          description: List of deposits
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/DepositResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/InternalError"

  # ──────────────────────────────────────────────
  #   Settlements
  # ──────────────────────────────────────────────
  /api/v1/settlements:
    get:
      operationId: listSettlements
      summary: List settlement batch history
      security: [{ BearerAuth: [] }, { CookieAuth: [] }]
      parameters:
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
          description: Maximum results (max 100)
        - name: offset
          in: query
          required: false
          schema:
            type: integer
            minimum: 0
            default: 0
          description: Number of results to skip
      responses:
        "200":
          description: Paginated list of settlement batches
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ListSettlementsResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/InternalError"

  # ──────────────────────────────────────────────
  #   Gateway
  # ──────────────────────────────────────────────
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: "ca_<key>"
      description: |
        API key authentication via `Authorization: Bearer ca_<key>`.

        Used for server-to-server and gateway proxy requests.
        Keys are prefixed with `ca_`, e.g. `ca_f8a2b1c3d4e5...`.
    CookieAuth:
      type: apiKey
      in: cookie
      name: session_token
      description: |
        OAuth session cookie. Set by the backend after a successful OAuth
        login (Google / GitHub). Used for browser dashboard requests.
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Conflict:
      description: Resource conflict
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    InternalError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
...
    CreateKeyRequest:
      type: object
      properties:
        label:
          type: string
          description: Human-readable label for the key
          example: My Production Key
        expires_at:
          type: string
          format: date-time
          nullable: true
          description: |
            Expiration timestamp. Defaults to 30 days from creation
            if not provided.
      required: [label]
    CreateKeyResponse:
      type: object
      properties:
        key:
          type: string
          description: The raw API key (shown once)
          example: ca_f8a2b1c3d4e5f6a7b8c9d0e1f2a3b4c5
        id:
          type: string
          format: uuid
          description: API key ID
        label:
          type: string
          description: Human-readable label
        status:
          type: string
          enum: [active, revoked, expired]
        created_at:
          type: string
          format: date-time
        expires_at:
          type: string
          format: date-time
          nullable: true
      required: [key, id, label, status, created_at]
    ListKeysItem:
      type: object
      properties:
        id:
          type: string
          format: uuid
        label:
          type: string
          nullable: true
        status:
          type: string
          enum: [active, revoked, expired]
        created_at:
          type: string
          format: date-time
        expires_at:
          type: string
          format: date-time
          nullable: true
      required: [id, label, status, created_at]
    RevokeKeyResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        label:
          type: string
        status:
          type: string
          enum: [active, revoked, expired]
        created_at:
          type: string
          format: date-time
        expires_at:
          type: string
          format: date-time
          nullable: true
      required: [id, label, status, created_at]

    # ── Auth ────────────────────────────────────
    UserResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
          description: User ID
        email:
          type: string
          format: email
          description: User email (from OAuth provider)
        created_at:
          type: string
          format: date-time
          description: Account creation timestamp
      required: [id, email, created_at]

    # ── Providers ───────────────────────────────
    CreateProviderRequest:
      type: object
      properties:
        name:
          type: string
          description: Provider name
          example: OpenAI
        base_url:
          type: string
          format: uri
          description: Upstream base URL
          example: https://api.openai.com
      required: [name, base_url]
    UpdateProviderRequest:
      type: object
      properties:
        name:
          type: string
          description: Provider name
        base_url:
          type: string
          format: uri
          description: Upstream base URL
      minProperties: 1
    UpdateProviderStatusRequest:
      type: object
      properties:
        status:
          type: string
          enum: [active, inactive, suspended]
          description: New provider status
      required: [status]
    Provider:
      type: object
      properties:
        id:
          type: string
          format: uuid
        owner_id:
          type: string
          format: uuid
        name:
          type: string
        base_url:
          type: string
          format: uri
        status:
          type: string
          enum: [active, inactive, suspended]
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
      required: [id, owner_id, name, base_url, status, created_at, updated_at]

    # ── Endpoints ───────────────────────────────
    CreateEndpointRequest:
      type: object
      properties:
        route:
          type: string
          description: Route path (e.g. `v1/chat/completions`)
          example: v1/chat/completions
        method:
          type: string
          description: HTTP method
          example: POST
        price_amount:
          type: string
          description: Price per request in the specified currency
          example: "0.01"
        currency:
          type: string
          enum: [XLM, USDC]
          default: XLM
          description: Billing currency (defaults to XLM)
        rate_limit:
          type: integer
          nullable: true
          description: Max requests per rate limit window
      required: [route, method, price_amount]
...
    UpdateEndpointStatusRequest:
      type: object
      properties:
        status:
          type: string
          enum: [active, inactive, draft]
          description: New endpoint status
      required: [status]
    Endpoint:
      type: object
      properties:
        id:
          type: string
          format: uuid
        provider_id:
          type: string
          format: uuid
        route:
          type: string
        method:
          type: string
        price_amount:
          type: string
          description: Price per request as a decimal string
        currency:
          type: string
          enum: [XLM, USDC]
        rate_limit:
          type: integer
          nullable: true
        status:
          type: string
          enum: [active, inactive, draft]
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
      required: [id, provider_id, route, method, price_amount, currency, status, created_at, updated_at]

    # ── Accounts ────────────────────────────────
...